requirements.txt
anthropic>=0.39.0
pytz>=2024.1
python-dotenv>=1.0.0